+++ isaaclab_arena/tests/test_arena_env_graph_spec.py
@@ -20,6 +20,8 @@ from isaaclab_arena.environments.graph_spec_utils import relation_class_for_spat
 from isaaclab_arena.relations.relations import AtPosition, IsAnchor, On, PositionLimits
 
 TEST_DATA_DIR = Path(__file__).parent / "test_data"
+_INIT_GRAPH = TEST_DATA_DIR / "pick_and_place_maple_table_init_env_graph.yaml"
+_FULL_GRAPH = TEST_DATA_DIR / "pick_and_place_maple_table_env_graph.yaml"
 
 
 def test_arena_env_graph_spec_loads_pick_and_place_yaml():
@@ -307,6 +309,67 @@ def test_arena_env_graph_spec_rejects_invalid_data():
         with pytest.raises(ValidationError) as exc_info:
             ArenaEnvGraphSpec.from_dict(data)
         assert error_match in str(exc_info.value), label
+
+
+# --- ArenaEnvGraphSpec.resolve_constraints: chaining a partially-wired graph into a full one ---
+
+
+def test_resolve_constraints_reproduces_groundtruth_full_graph():
+    """Resolving the partial init graph yields the hand-authored full graph (structurally)."""
+    resolved = ArenaEnvGraphSpec.from_yaml(_INIT_GRAPH, is_task_wiring_enabled=False).resolve_constraints()
+    groundtruth = ArenaEnvGraphSpec.from_yaml(_FULL_GRAPH)
+
+    # N tasks yield N+1 state specs, chained 0..N.
+    assert set(resolved.state_specs_by_id) == set(groundtruth.state_specs_by_id)
+    assert len(resolved.state_specs) == len(resolved.tasks) + 1
+
+    # Each state's spatial + task constraints match (keyed by id, so order is irrelevant).
+    for state_id, groundtruth_state in groundtruth.state_specs_by_id.items():
+        got = resolved.state_specs_by_id[state_id]
+        assert _spatial_by_id(got) == _spatial_by_id(groundtruth_state), f"spatial mismatch in {state_id}"
+        assert _task_constraints_by_id(got) == _task_constraints_by_id(
+            groundtruth_state
+        ), f"task mismatch in {state_id}"
+
+    # Tasks are wired into the chain identically.
+    assert set(resolved.tasks_by_id) == set(groundtruth.tasks_by_id)
+    for task_id, groundtruth_task in groundtruth.tasks_by_id.items():
+        got = resolved.tasks_by_id[task_id]
+        assert got.type == groundtruth_task.type
+        assert got.initial_state_spec_id == groundtruth_task.initial_state_spec_id
+        assert got.success_state_spec_id == groundtruth_task.success_state_spec_id
+        assert got.task_args == groundtruth_task.task_args
+
+
+def test_unresolved_graph_is_not_directly_loadable():
+    """The init graph leaves tasks unwired, so the strict loader rejects it -- motivating resolve_constraints."""
+    with pytest.raises(ValidationError):
+        ArenaEnvGraphSpec.from_yaml(_INIT_GRAPH)
+
+
+def test_chain_wires_each_success_state_as_next_initial_state():
+    """task[i].success_state_spec_id == task[i+1].initial_state_spec_id (a single chain)."""
+    resolved = ArenaEnvGraphSpec.from_yaml(_INIT_GRAPH, is_task_wiring_enabled=False).resolve_constraints()
+    ordered = sorted(resolved.tasks_by_id.values(), key=lambda t: t.id)
+    for earlier, later in zip(ordered, ordered[1:]):
+        assert earlier.success_state_spec_id == later.initial_state_spec_id
+
+
+def test_task_without_a_transition_is_rejected():
+    """A task whose class declares no success_state_transition fails loudly rather than silently skipping."""
+    spec = ArenaEnvGraphSpec.from_yaml(_INIT_GRAPH, is_task_wiring_enabled=False)
+    spec.tasks[0].type = "NoTask"  # registered, but declares no transition
+    with pytest.raises(NotImplementedError, match="success_state_transition not implemented"):
+        spec.resolve_constraints()
+
+
+def _spatial_by_id(state: ArenaEnvGraphStateSpec) -> dict[str, tuple]:
+    """Project a state's spatial constraints to id -> (type, parent, child, params), order-insensitive."""
+    return {c.id: (c.type, c.parent, c.child, c.params) for c in state.spatial_constraints}
+
+
+def _task_constraints_by_id(state: ArenaEnvGraphStateSpec) -> dict[str, tuple]:
+    return {c.id: (c.type, c.parent, c.child, c.params) for c in state.task_constraints}
 
 
 def _minimal_env_graph_data():
